+++ setup/install-docker.sh
@@ -14,11 +14,10 @@ fi
 
 # Generate and build OpenAPI JavaScript client
 docker compose run neucore_java /app/frontend/openapi.sh
-docker compose exec neucore_node npm install --prefix /app/frontend/neucore-js-client
+docker compose exec neucore_node npm install --ignore-scripts --prefix /app/frontend/neucore-js-client
 docker compose exec neucore_node npm run build --prefix /app/frontend/neucore-js-client
 
 # Build frontend
-docker compose exec neucore_node npm i file:neucore-js-client
 docker compose exec neucore_node npm ci
 if [[ $1 = prod ]]; then
     docker compose exec neucore_node npm run build