+++ setup/install-podman.sh
@@ -14,11 +14,10 @@ fi
 
 # Generate and build OpenAPI JavaScript client
 podman compose run neucore_java /app/frontend/openapi.sh
-podman compose exec neucore_node npm install --prefix /app/frontend/neucore-js-client
+podman compose exec neucore_node npm install --ignore-scripts --prefix /app/frontend/neucore-js-client
 podman compose exec neucore_node npm run build --prefix /app/frontend/neucore-js-client
 
 # Build frontend
-podman compose exec neucore_node npm i file:neucore-js-client
 podman compose exec neucore_node npm ci
 if [[ $1 = prod ]]; then
     podman compose exec neucore_node npm run build